gilang

+++ application/modules/newscategory/controllers/RequestController.php
@@ -151,15 +151,25 @@ class Newscategory_RequestController extends Zend_Controller_Action
 	{
 		if($this->getRequest()->isPost() && $this->getRequest()->isXmlHttpRequest() && isset($this->_posts[$this->_pk])) {
 			if($this->_model->isExistByKey($this->_pk, $this->_posts[$this->_pk])) {
-	
+				
 				try {
-					$this->_model->delete($this->_model->getAdapter()->quoteInto($this->_pk . ' = ?', $this->_posts[$this->_pk]));
+					$researchModel = new Application_Model_News();
+					$q = $researchModel->select()
+					->where('NEWS_CATEGORY_ID = ?', $this->_posts[$this->_pk]);
+					$count = $q->query()->fetchAll();
+					if(count($count) == 0) {
+						$this->_model->delete($this->_model->getAdapter()->quoteInto($this->_pk . ' = ?', $this->_posts[$this->_pk]));
+					} else {
+						$this->_error_code = 202;
+						$this->_error_message = MyIndo_Tools_Error::getErrorMessage($this->_error_code);
+						$this->_success = false;
+					}
 				}catch(Exception $e) {
 					$this->_error_code = $e->getCode();
 					$this->_error_message = $e->getMessage();
 					$this->_success = false;
 				}
-	
+				
 			} else {
 				$this->_error_code = 101;
 				$this->_error_message = MyIndo_Tools_Error::getErrorMessage($this->_error_code);
